pyproject.toml
[build-system]
requires = ['setuptools>=61.0.0', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = "python-service-builder"
version = "1.0.0"
authors = [{ name = "David Harcombe", email = "david.harcombe@gmail.com" }]
description = "Helper library for easier creation of Google services."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
 "Programming Language :: Python :: 3",
 "Operating System :: OS Independent",
]
license = { text = "Apache 2.0" }
dependencies = [
 'absl-py>=0.12.0',
 'dataclasses-json>=0.5.2',
 'gcs-oauth2-boto-plugin>=2.7',
 'google-api-core>=1.26.1',
 'google-api-python-client>=2.0.2',
 'google-auth-httplib2>=0.1.0',
 'google-auth-oauthlib>=0.4.3',
 'google-auth>=1.28.0',
 'google-reauth>=0.1.1',
 'googleapis-common-protos>=1.53.0',
 'httplib2>=0.19.0',
 'immutabledict>=2.2.1',
 'oauth2client>=4.1.3',
 'oauthlib>=3.1.0',
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["service_framework*"]
exclude = ["*_test.py"]

[project.urls]
"Homepage" = "https://github.com/google/python-service-builder"
"Bug Tracker" = "https://github.com/google/python-service-builder/issues"
